(nhncs) Disable Terminal 'resume'
Use Bash history instead of Terminal resume:

https://stackoverflow.com/questions/32418438/how-can-i-disable-bash-sessions-in-os-x-el-capitan/46248280

"A seemingly new feature in OS X El Capitan (10.11 Beta) is Bash
sessions (Terminal sessions). I now have a ~/.bash_sessions directory with
history files, and my HISTFILE and HISTIGNORE envars are being overridden. How
can I disable all of this functionality?"

https://stackoverflow.com/a/46248280

"Finally, and this relates specifically to the OP's question, it will restore the history in each window/tab if the following conditions hold:

SHELL_SESSION_HISTORY is unset or set to 1
shopt histappend is not set
HISTTIMEFORMAT is not set
If SHELL_SESSION_HISTORY is explicitly set to 1 in .bashrc, the last two requirements are overridden, that is, shopt histappend or HISTTIMEFORMAT could be set."